requirements.txt
edge-tts>=6.1.9
pygame>=2.5.2
flask>=2.0.0
tiktoken>=0.5.0
dirtyjson>=1.0.7
requests>=2.31.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.3.16
langchain-core>=0.3.58
langchain-community>=0.0.16
langchain-google-genai>=0.0.5
langchain-ollama>=0.0.1
python-dotenv==1.0.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
lxml_html_clean>=0.4.0
openai>=1.0.0
html2text>=2020.1.16
selenium>=4.0.0
webdriver-manager>=4.0.0
tavily-python>=0.3.0 